Return HTTP 400 Bad Request for bad request body
Amended helper method HandleBodyDecodedValidation to return
http.StatusBadRequest on error & updated relevant unit test checks to match.

One unit test check which shared the same name (in
attachment_resource_test) was actually performing a slightly different
check, so I have amended the name of the 'Convey' scope to reflect this
(and as I was touching it I also updated the test to use the real
helperService & added an error message check based on its current
behaviour).