Additional test plus minor fix [IEP-753]
Adding one more test and a minor fix that was revealed by trying to add
it - test initially failed because mtest objected that a session was
left open, due to return in the middle of iterating through a cursor.
Added a call to Close on the cursor before return and tests now pass.